requirements.txt
# Core dependencies
pillow>=10.0.0
customtkinter>=5.2.0
pywin32>=306; sys_platform == 'win32'
winshell>=0.6; sys_platform == 'win32'

# Cross-platform support
pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'
python-xlib>=0.33; sys_platform == 'linux'

# Modern UI components
darkdetect>=0.8.0
tkinterdnd2>=0.3.0

# Performance and async
aiofiles>=23.0.0
asyncio>=3.4.3

# Development dependencies (optional)
black>=23.0.0
pylint>=2.17.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Build tools
pyinstaller>=6.0.0
auto-py-to-exe>=2.40.0

# Additional utilities
requests>=2.31.0
psutil>=5.9.0
colorama>=0.4.6